Moving all API call functions to one file - useAPIUtil.js
- Fixing issue with the PingBackend function which didn't have a '/search' postfix string after the environment variable URL
- Updating test file with updated imports.
