+++ src/api/useApiUtil.js
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 
-const useApiUtil = (term) => {
+export const useApiUtil = (term) => {
     const [data, setData] = useState(null);
 
     useEffect(() => {
@@ -41,4 +41,25 @@ const useApiUtil = (term) => {
     return { data };
 }
 
-export default useApiUtil;
+export const PingBackend = async () => {
+    try{
+        const response = await fetch(process.env.REACT_APP_BACKEND_URL + "/search", {
+            method: "POST",
+            headers: {
+                'Content-Type':'application/json'
+            },
+            body: JSON.stringify({
+                query: ""
+            }),
+        });
+
+        if(response.ok) {
+            console.log("Active Server Connection");
+        } else {
+            throw new Error("Request failed");
+        }
+        
+    } catch(error) {
+        console.error("Error: ", error);
+    }
+}
