+++ src/pages/home/Home.js
@@ -3,7 +3,7 @@ import styles from "../home/Home.module.css";
 import Header from '../../components/Header';
 import CFIALogo from '../../components/CFIALogo';
 import { useState, useEffect } from "react";
-import { PingBackend } from '../../api/api';
+import { PingBackend } from '../../api/useApiUtil';
 
 const Home = () => {
     const [alertMessage, setAlertMessage] = useState(""); // State variable for the alert message
@@ -25,7 +25,7 @@ const Home = () => {
             setIsError(true);
             setAlertMessage("Warning: Backend URL is not set, frontend is misconfigured.")
           }
-    });
+    }, []);
 
     return (
         <div className={styles.layout}>
